pyproject.toml
[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # длинные строки в миграциях
    "B008",  # Depends() в defaults (FastAPI)
    "E712",  # == True/False (SQLAlchemy требует)
    "B904",  # raise from (стиль)
    "E402",  # import not at top (иногда необходимо)
]

[tool.ruff.lint.isort]
known-first-party = ["api", "integrations", "cad-cam", "shared"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "furniture-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",
	"sqlalchemy>=2.0.32",
	"asyncpg>=0.29.0",
	"alembic>=1.13.2",
	"redis>=5.0.7",
	"boto3>=1.34.150",
	"botocore>=1.34.150",
	"pydantic-settings>=2.4.0",
	"aiohttp>=3.10.0",
	"numpy>=1.26.0",
	"scipy>=1.13.0",
	"pgvector>=0.4.1",
	"pandas>=2.2.0",
	"openpyxl>=3.1.0",
	"typer>=0.9.0",
	"psycopg[binary]>=3.1.0",
	"ezdxf>=1.4.3",
	"pillow>=12.1.0",
	"rectpack>=0.2.2",
	"pyjwt>=2.8.0",
	"passlib[bcrypt]>=1.7.4",
	"email-validator>=2.3.0",
	"pymupdf>=1.26.7",
	"openai>=2.15.0",
	"python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]
